Tool

+++ Tool/Z.Tool.PrusateGen/Gen.cs
@@ -4,13 +4,13 @@ class Gen : Any
 {
     protected virtual ReadResult ReadResult { get; set; }
 
-    public virtual int Execute()
+    public virtual long Execute()
     {
         Read read;
         read = new Read();
         read.Init();
 
-        int o;
+        long o;
         o = read.Execute();
 
         if (!(o == 0))
